requirements.txt
streamlit
pandas
duckdb
langchain
langchain-groq
python-dotenv